requirements.txt
Flask==2.0.1
pylint==2.8.2
python-dotenv==0.17.1
requests==2.25.1
pytest==6.2.4